+++ python/python/glide/async_commands/core.py
@@ -16,7 +16,7 @@ from typing import (
     get_args,
 )
 
-from glide.async_commands.command_args import Limit, OrderBy
+from glide.async_commands.command_args import Limit, ListDirection, OrderBy
 from glide.async_commands.sorted_set import (
     AggregationType,
     InfBound,
@@ -1493,6 +1493,101 @@ class CoreCommands(Protocol):
             int,
             await self._execute_command(
                 RequestType.LInsert, [key, position.value, pivot, element]
+            ),
+        )
+
+    async def lmove(
+        self,
+        source: str,
+        destination: str,
+        where_from: ListDirection,
+        where_to: ListDirection,
+    ) -> Optional[str]:
+        """
+        Atomically pops and removes the left/right-most element to the list stored at `source`
+        depending on `where_from`, and pushes the element at the first/last element of the list
+        stored at `destination` depending on `where_to`.
+
+        When in cluster mode, both `source` and `destination` must map to the same hash slot.
+
+        See https://valkey.io/commands/lmove/ for details.
+
+        Args:
+            source (str): The key to the source list.
+            destination (str): The key to the destination list.
+            where_from (ListDirection): The direction to remove the element from (`ListDirection.LEFT` or `ListDirection.RIGHT`).
+            where_to (ListDirection): The direction to add the element to (`ListDirection.LEFT` or `ListDirection.RIGHT`).
+
+        Returns:
+            Optional[str]: The popped element, or None if `source` does not exist.
+
+        Examples:
+            >>> client.lpush("testKey1", ["two", "one"])
+            >>> client.lpush("testKey2", ["four", "three"])
+            >>> await client.lmove("testKey1", "testKey2", ListDirection.LEFT, ListDirection.LEFT)
+            "one"
+            >>> updated_array1 = await client.lrange("testKey1", 0, -1)
+            ["two"]
+            >>> await client.lrange("testKey2", 0, -1)
+            ["one", "three", "four"]
+
+        Since: Redis version 6.2.0.
+        """
+        return cast(
+            Optional[str],
+            await self._execute_command(
+                RequestType.LMove,
+                [source, destination, where_from.value, where_to.value],
+            ),
+        )
+
+    async def blmove(
+        self,
+        source: str,
+        destination: str,
+        where_from: ListDirection,
+        where_to: ListDirection,
+        timeout: float,
+    ) -> Optional[str]:
+        """
+        Blocks the connection until it pops atomically and removes the left/right-most element to the
+        list stored at `source` depending on `where_from`, and pushes the element at the first/last element
+        of the list stored at `destination` depending on `where_to`.
+        `BLMOVE` is the blocking variant of `LMOVE`.
+
+        Notes:
+            1. When in cluster mode, both `source` and `destination` must map to the same hash slot.
+            2. `BLMOVE` is a client blocking command, see https://github.com/aws/glide-for-redis/wiki/General-Concepts#blocking-commands for more details and best practices.
+
+        See https://valkey.io/commands/blmove/ for details.
+
+        Args:
+            source (str): The key to the source list.
+            destination (str): The key to the destination list.
+            where_from (ListDirection): The direction to remove the element from (`ListDirection.LEFT` or `ListDirection.RIGHT`).
+            where_to (ListDirection): The direction to add the element to (`ListDirection.LEFT` or `ListDirection.RIGHT`).
+            timeout (float): The number of seconds to wait for a blocking operation to complete. A value of `0` will block indefinitely.
+
+        Returns:
+            Optional[str]: The popped element, or None if `source` does not exist or if the operation timed-out.
+
+        Examples:
+            >>> await client.lpush("testKey1", ["two", "one"])
+            >>> await client.lpush("testKey2", ["four", "three"])
+            >>> await client.blmove("testKey1", "testKey2", ListDirection.LEFT, ListDirection.LEFT, 0.1)
+            "one"
+            >>> await client.lrange("testKey1", 0, -1)
+            ["two"]
+            >>> updated_array2 = await client.lrange("testKey2", 0, -1)
+            ["one", "three", "four"]
+
+        Since: Redis version 6.2.0.
+        """
+        return cast(
+            Optional[str],
+            await self._execute_command(
+                RequestType.BLMove,
+                [source, destination, where_from.value, where_to.value, str(timeout)],
             ),
         )